Consolidate Footnote components and hook into same folder
