+++ src/components/ComponentFactory.js
@@ -25,8 +25,8 @@ import Extract from './Extract';
 import Field from './Field';
 import FieldList from './FieldList';
 import Figure from './Figure';
-import Footnote from './Footnote';
-import FootnoteReference from './FootnoteReference';
+import Footnote from './Footnote/Footnote';
+import FootnoteReference from './Footnote/FootnoteReference';
 import Glossary from './Glossary';
 import GuideNext from './GuideNext/GuideNext';
 import Heading from './Heading';
