+++ src/components/Footnote/FootnoteReference.js
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import PropTypes from 'prop-types';
-import { getNestedValue } from '../utils/get-nested-value';
 import FootnoteContext from './footnote-context';
+import { getNestedValue } from '../../utils/get-nested-value';
 
 const FootnoteReference = ({ nodeData: { id, refname } }) => {
   const { footnotes } = useContext(FootnoteContext);
