+++ tests/unit/FootnoteReference.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import FootnoteReference from '../../src/components/FootnoteReference';
+import FootnoteReference from '../../src/components/Footnote/FootnoteReference';
 
 // data for this component
 import mockData from './data/FootnoteReference.test.json';
